avoid a global function
