+++ resources/js/tests/Globals.test.js
@@ -76,63 +76,3 @@ test('it_can_set_value_to_data_using_dotted_path', () => {
 
     expect(data).toStrictEqual(expected);
 });
-
-test('it_can_delete_value_from_data_using_dotted_path', () => {
-    let data = {
-        foo: 'alpha',
-        nested: {
-            foo: 'beta',
-            array_value: [
-                {foo: 'charlie'},
-                {foo: 'delta'},
-            ],
-        },
-    };
-
-    data_delete(data, 'nested.array_value.1.foo');
-
-    expect(data).toStrictEqual({
-        foo: 'alpha',
-        nested: {
-            foo: 'beta',
-            array_value: [
-                {foo: 'charlie'},
-                {},
-            ],
-        },
-    });
-
-    data_delete(data, 'foo');
-
-    expect(data).toStrictEqual({
-        nested: {
-            foo: 'beta',
-            array_value: [
-                {foo: 'charlie'},
-                {},
-            ],
-        },
-    });
-
-    data_delete(data, 'nested.foo');
-
-    expect(data).toStrictEqual({
-        nested: {
-            array_value: [
-                {foo: 'charlie'},
-                {},
-            ],
-        },
-    });
-});
-
-
-test('it throws error when trying to delete an array item', () => {
-    let data = {
-        foo: ['alfa', 'bravo', 'charlie'],
-    };
-
-    expect(() => {
-        data_delete(data, 'foo.1');
-    }).toThrow('Cannot delete an array key');
-});
